break infinite loop when updating elements with FLNK to item record
When updating records from the driver, disable FLNK unless SCAN="I/O Intr".
Otherwise if the (output) record is a structure element and the FLNK causes
the parent opcuaItem record to process, an inifnite processing loop is created.
Also, the user probably only expects the FLNK to process when actively writing
to the output record, not when updating it from the server.

Fixes issue #145.